Add files via upload
In our challenge, we are asked to explain the problem of greenhouse gases and the carbon consumption rate on Earth.  We developed a website that gathers all the information related to gases on the planet, especially carbon dioxide.  Our website uses maps and explanations to illustrate carbon dioxide levels on Earth, providing links to sources for each explanation.  The content is organized around topics related to global warming and carbon dioxide consumption rates on Earth.  That was the main challenge, but we added some features.  We have implemented a login system where users can log in using their email and password.  After logging in, users can view a list of items that contribute to the rise in Earth's temperatures, such as paper, plastic, and glass.  The user collects a certain quantity of these materials, enters the collected quantity, and points are calculated for each material.  These points can be exchanged for discounts or cash in some restaurants or stores.  In addition, we created a tool to calculate monthly household energy consumption, allowing users to see the amount of carbon dioxide emitted per kilowatt-hour of electricity produced.
This calculation is based on the average monthly consumption for a family of five, which is approximately 400 kilowatt-hours of electricity.  We also introduced a feature to calculate the energy consumption for each device